requirements.txt
streamlit
opencv-python-headless
numpy
pillow